requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.3
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.6
Faker==19.6.1
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
